fix: handle error in executeMainThread
